Remove CountdownView
We moved the countdown to the overlay instead over the button.
