+++ PicIt/View/CameraCaptureButtonView.swift
@@ -2,6 +2,27 @@
 
 import SwiftUI
 
+extension CountdownState {
+    
+    func cameraButtonColor() -> Color {
+        switch self {
+        case .stopped:
+            return .gray
+        case .inProgress:
+            return .yellow
+        case .triggering:
+            return .green
+        case .complete:
+            return .mint
+        case .undefined:
+            return .black
+        case .ready:
+            return .white
+        }
+    }
+    
+}
+
 struct CameraCaptureButton: View {
     static let log = PicItSelfLog<CameraCaptureButton>.get()
     
@@ -15,8 +36,16 @@ struct CameraCaptureButton: View {
     //    ok for CameraCaptureModel([weak]? cameraAction, countdown?)
     
     var buttonView: some View {
-        HStack {
-            CountdownView(countdownTimer: countdownTime, countdownState: countdownState)
+        ZStack {
+            Circle()
+                .foregroundColor(countdownState.cameraButtonColor())
+                .frame(width: 80, height: 80, alignment: .center)
+                .overlay(
+                    Circle()
+                        .stroke(Color.black.opacity(0.8), lineWidth: 2)
+                        .frame(width: 65, height: 65, alignment: .center)
+                )
+
             mediaMode.indicatorView()
         }
     }
